pyproject.toml
[tool.poetry]
name = "graphanalytica"
version = "0.1.0"
description = "A frontend for pyrthomas"
authors = ["Talha Junaid <talhajunaidd@gmail.com>"]

#packages = [
#    { include = "pyrthomas", from="localpath/pyrthomas" },
#]


[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
djangorestframework = "^3.12.4"
networkx = "^2.5.1"
whitenoise = "^5.2.0"
gunicorn = "^20.1.0"
pyrthomas = "^0.0.10"


[tool.poetry.dev-dependencies]
honcho = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

